live-api/pyproject.toml
[project]
name = "live-api"
version = "0.1.0"
description = "Real-Time Live AI Assistant using Gemini Live API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.52.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pyaudio>=0.2.14",
    "lmnr>=0.4.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
]
